requirements.txt
pdf2image
pytesseract
Pillow
openai
python-dotenv
tqdm